Abstract out frame delay rewriter.
