+++ coil-gif/src/main/java/coil/decode/GifDecoder.kt
@@ -5,8 +5,6 @@ package coil.decode
 import android.graphics.Bitmap
 import android.graphics.Movie
 import android.os.Build.VERSION.SDK_INT
-import androidx.annotation.CheckResult
-import androidx.annotation.VisibleForTesting
 import coil.bitmap.BitmapPool
 import coil.drawable.MovieDrawable
 import coil.request.animatedTransformation
@@ -16,9 +14,7 @@ import coil.request.repeatCount
 import coil.size.Size
 import coil.util.animatable2CompatCallbackOf
 import coil.util.isHardware
-import okio.Buffer
 import okio.BufferedSource
-import okio.ByteString.Companion.decodeHex
 import okio.buffer
 
 /**
@@ -30,8 +26,10 @@ import okio.buffer
  *  it is below a threshold. See https://github.com/coil-kt/coil/issues/540 for more info.
  */
 class GifDecoder @JvmOverloads constructor(
-    private val enforceMinimumFrameDelay: Boolean = false
+    enforceMinimumFrameDelay: Boolean = false
 ) : Decoder {
+
+    private val frameDelayRewriter = FrameDelayRewriter(enforceMinimumFrameDelay)
 
     override fun handles(source: BufferedSource, mimeType: String?) = DecodeUtils.isGif(source)
 
@@ -41,12 +39,7 @@ class GifDecoder @JvmOverloads constructor(
         size: Size,
         options: Options
     ): DecodeResult = withInterruptibleSource(source) { interruptibleSource ->
-        val bufferedSource = if (enforceMinimumFrameDelay) {
-            rewriteFrameDelay(interruptibleSource.buffer())
-        } else {
-            interruptibleSource.buffer()
-        }
-
+        val bufferedSource = frameDelayRewriter.rewriteFrameDelay(interruptibleSource.buffer())
         val movie: Movie? = bufferedSource.use {
             if (SDK_INT >= 19) {
                 Movie.decodeStream(it.inputStream())
@@ -88,50 +81,7 @@ class GifDecoder @JvmOverloads constructor(
         )
     }
 
-    /** Rewrite the frame delay in each graphics control block if it's below a threshold. */
-    @CheckResult
-    @VisibleForTesting
-    internal fun rewriteFrameDelay(source: BufferedSource) = source.use {
-        val buffer = Buffer()
-        var index = 0L
-
-        // Search through the buffer and rewrite any frame delays below the threshold.
-        while (true) {
-            val frameDelayStartMarkerIndex = source.indexOf(FRAME_DELAY_START_MARKER, index)
-            if (frameDelayStartMarkerIndex == -1L) break
-
-            // Read up until the end of the frame delay start marker.
-            index = frameDelayStartMarkerIndex + FRAME_DELAY_START_MARKER.size
-            source.read(buffer, index - buffer.size)
-
-            // Check that the frame delay end marker is present, else this is a false positive.
-            if (!source.request(5) || source.buffer[4].toInt() != 0) continue
-
-            // Rewrite the frame delay if it is below the threshold.
-            if (source.buffer[2].toInt() < MINIMUM_FRAME_DELAY) {
-                buffer.writeByte(source.buffer[0].toInt())
-                buffer.writeByte(0)
-                buffer.writeByte(DEFAULT_FRAME_DELAY)
-                source.skip(3)
-                index += 3
-            }
-        }
-
-        // Write the rest of the source and return the buffer.
-        return@use buffer.apply { source.readAll(this) }
-    }
-
     companion object {
-        // The Graphics Control Extension block is guaranteed to match the following hexadecimal sequence:
-        // 00 21 F9 04 XX FD FD XX 00
-        // - FD is the frame delay value
-        // - XX matches any byte value
-        // https://www.matthewflickinger.com/lab/whatsinagif/images/graphic_control_ext.gif
-        private val FRAME_DELAY_START_MARKER = "0021F904".decodeHex()
-
-        private const val MINIMUM_FRAME_DELAY = 2
-        private const val DEFAULT_FRAME_DELAY = 10
-
         const val REPEAT_COUNT_KEY = "coil#repeat_count"
         const val ANIMATED_TRANSFORMATION_KEY = "coil#animated_transformation"
         const val ANIMATION_START_CALLBACK_KEY = "coil#animation_start_callback"
