+++ coil-sample/src/main/java/coil/sample/Application.kt
@@ -27,10 +27,10 @@ class Application : MultiDexApplication(), ImageLoaderFactory {
             .crossfade(true) // Show a short crossfade when loading images from network or disk.
             .componentRegistry {
                 // GIFs
-                if (false && SDK_INT >= 28) {
-                    add(ImageDecoderDecoder(this@Application))
+                if (SDK_INT >= 28) {
+                    add(ImageDecoderDecoder(this@Application, enforceMinimumFrameDelay = true))
                 } else {
-                    add(GifDecoder(enforceMinimumFrameDelay = true))
+                    add(GifDecoder(enforceMinimumFrameDelay = false))
                 }
 
                 // SVGs
